Rename for minimal confusion.
