+++ include/hip/hcc_detail/math_fwd.h
